+++ pandora-common/src/assets/definitions.ts
@@ -1,18 +1,15 @@
+import { Immutable } from 'immer';
 import { z } from 'zod';
+import type { AssetPreferenceType } from '../character';
 import type { RoomBackgroundData } from '../space/room';
+import type { Satisfies } from '../utility';
 import { HexRGBAColorString } from '../validation';
-import type { AppearanceArmPose, AppearancePose } from './state/characterState';
-import type { BoneDefinitionCompressed, BoneName, BoneType, CharacterView, Condition, Coordinates, LayerImageOverride, LegsPose } from './graphics';
+import type { AssetId } from './base';
+import type { BoneDefinitionCompressed, BoneName, CharacterView, Condition, Coordinates, LayerImageOverride, LegsPose } from './graphics';
 import type { AssetModuleDefinition } from './modules';
 import type { AssetLockProperties, AssetProperties } from './properties';
-import type { Satisfies } from '../utility';
-import { Immutable, freeze, produce } from 'immer';
-import type { AssetManager } from './assetManager';
-import _ from 'lodash';
-import { BONE_MAX, BONE_MIN } from './appearance';
 import type { RoomDeviceProperties } from './roomDeviceProperties';
-import type { AssetPreferenceType } from '../character';
-import type { AssetId } from './base';
+import type { AppearanceArmPose, AssetsPosePreset, AssetsPosePresets } from './state/characterStatePose';
 
 // Each asset must have a size (bodyparts and only bodyparts have `bodypart` size)
 // The size is used to make sure you cannot infinitely recurse storing items into one another
@@ -379,96 +376,6 @@ export interface AssetBodyPart {
 	/** If changes to this bodypart are not considered as "body changes", lessening restrictions */
 	adjustable: boolean;
 }
-
-export type PartialAppearancePose<Bones extends BoneName = BoneName> = {
-	bones?: Partial<Record<Bones, number>>;
-	arms?: Partial<AppearanceArmPose>;
-	leftArm?: Partial<AppearanceArmPose>;
-	rightArm?: Partial<AppearanceArmPose>;
-	legs?: LegsPose;
-	view?: CharacterView;
-};
-
-export function MergePartialAppearancePoses(base: Immutable<PartialAppearancePose>, extend?: Immutable<PartialAppearancePose>): PartialAppearancePose {
-	if (extend == null)
-		return base;
-
-	return {
-		bones: { ...base.bones, ...extend.bones },
-		arms: { ...base.arms, ...extend.arms },
-		leftArm: { ...base.leftArm, ...extend.leftArm },
-		rightArm: { ...base.rightArm, ...extend.rightArm },
-		legs: base.legs ?? extend.legs,
-		view: base.view ?? extend.view,
-	};
-}
-
-export function ProduceAppearancePose(
-	basePose: Immutable<AppearancePose>,
-	{
-		assetManager,
-		boneTypeFilter,
-		missingBonesAsZero = false,
-	}: {
-		assetManager: AssetManager;
-		boneTypeFilter?: BoneType;
-		/** @default false */
-		missingBonesAsZero?: boolean;
-	},
-	...changes: [(PartialAppearancePose | AssetsPosePreset), ...(PartialAppearancePose | AssetsPosePreset)[]]
-): Immutable<AppearancePose> {
-	const pose = changes.reduce(MergePartialAppearancePoses);
-
-	return produce(basePose, (draft) => {
-		// Update view
-		if (pose.view != null) {
-			draft.view = pose.view;
-		}
-
-		// Update arms
-		{
-			const leftArm = { ...basePose.leftArm, ...pose.arms, ...pose.leftArm };
-			const rightArm = { ...basePose.rightArm, ...pose.arms, ...pose.rightArm };
-			const armsChanged =
-				!_.isEqual(basePose.leftArm, leftArm) ||
-				!_.isEqual(basePose.rightArm, rightArm);
-
-			if (armsChanged) {
-				draft.leftArm = freeze(leftArm, true);
-				draft.rightArm = freeze(rightArm, true);
-			}
-		}
-
-		// Update legs
-		if (pose.legs != null) {
-			draft.legs = pose.legs;
-		}
-
-		// Update bones
-		if (pose.bones != null) {
-			for (const bone of assetManager.getAllBones()) {
-				const newValue = pose.bones[bone.name];
-
-				if (boneTypeFilter !== undefined && bone.type !== boneTypeFilter)
-					continue;
-				if (!missingBonesAsZero && newValue == null)
-					continue;
-
-				draft.bones[bone.name] = (newValue != null && Number.isInteger(newValue)) ? _.clamp(newValue, BONE_MIN, BONE_MAX) : 0;
-			}
-		}
-	});
-}
-
-export type AssetsPosePreset<Bones extends BoneName = BoneName> = PartialAppearancePose<Bones> & {
-	name: string;
-	optional?: PartialAppearancePose<Bones>;
-};
-
-export type AssetsPosePresets<Bones extends BoneName = BoneName> = {
-	category: string;
-	poses: AssetsPosePreset<Bones>[];
-}[];
 
 export type AssetAttributeDefinition<A extends AssetDefinitionExtraArgs = AssetDefinitionExtraArgs> = {
 	name: string;
