+++ pandora-common/src/assets/state/_index.ts
@@ -1,3 +1,5 @@
 export * from './characterState';
+export * from './characterStatePose';
+export * from './characterStateTypes';
 export * from './globalState';
 export * from './roomState';
