+++ pandora-common/src/character/index.ts
@@ -3,4 +3,5 @@ export * from './characterData';
 export * from './characterTypes';
 export * from './pronouns';
 export * from './restrictionsManager';
+export * from './restrictionTypes';
 export * from './speech';
