+++ pandora-common/src/gameLogic/character/character.ts
@@ -1,12 +1,12 @@
-import { CharacterId, CharacterRestrictionsManager, ICharacterMinimalData } from '../../character';
 import { AccountId } from '../../account';
-import { TypedEventEmitter } from '../../event';
-import { InteractionSubsystem } from '../interactions/interactionSubsystem';
 import { AssetFrameworkCharacterState, CharacterAppearance } from '../../assets';
+import { CharacterId, CharacterRestrictionsManager, ICharacterMinimalData } from '../../character';
+import { TypedEventEmitter } from '../../event';
+import type { ActionSpaceContext } from '../../space/space';
 import { Assert } from '../../utility';
-import { ActionSpaceContext } from '../../space/space';
-import { GameLogicPermission, IPermissionProvider, PermissionGroup } from '../permissions';
 import { AssetPreferencesSubsystem } from '../assetPreferences';
+import { InteractionSubsystem } from '../interactions/interactionSubsystem';
+import { GameLogicPermission, IPermissionProvider, PermissionGroup } from '../permissions';
 
 export type GameLogicCharacterEvents = {
 	dataChanged: 'interactions' | 'assetPreferences';
