+++ pandora-common/src/networking/connection.ts
@@ -1,13 +1,12 @@
-import { IsObject } from '../validation';
-import { DEFAULT_ACK_TIMEOUT } from './config';
-import { GetLogger, Logger } from '../logging';
-import type { SocketInterfaceDefinition, SocketInterfaceOneshotMessages, SocketInterfaceRequest, SocketInterfaceRespondedMessages, SocketInterfaceResponse } from './helpers';
-import type { IServerSocket, ServerRoom } from './room';
-import { MESSAGE_HANDLER_DEBUG_ALL, MESSAGE_HANDLER_DEBUG_MESSAGES } from './config';
-import { EmitterWithAck, IncomingSocket, MessageCallback, MockConnectionSocket } from './socket';
-import { BadMessageError, IMessageHandler } from './message_handler';
-import { AssertNotNullable } from '../utility';
 import { nanoid } from 'nanoid';
+import { GetLogger, Logger } from '../logging';
+import { AssertNotNullable } from '../utility';
+import { IsObject } from '../validation';
+import { DEFAULT_ACK_TIMEOUT, MESSAGE_HANDLER_DEBUG_ALL, MESSAGE_HANDLER_DEBUG_MESSAGES } from './config';
+import type { SocketInterfaceDefinition, SocketInterfaceOneshotMessages, SocketInterfaceRequest, SocketInterfaceRespondedMessages, SocketInterfaceResponse } from './helpers';
+import { BadMessageError, IMessageHandler } from './message_handler';
+import type { IServerSocket, ServerRoom } from './room';
+import { EmitterWithAck, IncomingSocket, MessageCallback, MockConnectionSocket } from './socket';
 
 export interface IConnectionBase<OutboundT extends SocketInterfaceDefinition> {
 	/**
